feat(web): guard against missing or invalid links submissions
Limit ourselves to just testing for a present, valid ShopBack link.
As we are unable to manipulate Chrome's `fetch()` to manipulate the
User-Agent header (now deemed legal [^1]), we cannot have the frontend
attempt to parse metadata.

TODO:
See if we can at least have Firefox do the metadata validation

References:
[^1]: https://bugs.chromium.org/p/chromium/issues/detail?id=571722